Added 'pact' field and constructor in Knight class
The Knight class has been extended with a new field called 'pact'. This field is associated with a getter and a setter. Besides, a new constructor has been formulated that includes 'pact' as an attribute. The default constructor's implementation has been modified to call the parameterized constructor with default values. Supported with the translation of certain comments from French to English to improve readability.