Add detailed comments and improve Knight and Mirlky classes
In both Knight and Mirlky classes, added detailed comments to methods as well as getter and setter methods for class properties. Improved Knight class by changing the displayStats method to accept a Player object and sending the stats as player messages. Also, a jar artifact configuration was added in the .idea folder.